inline use-once snippets
