+++ packages/website/src/content/03-03-disclaimer.tsx
@@ -4,19 +4,57 @@ import { PageContent } from "../components/PageContent";
 import { Note } from "../components/Note";
 import { Disclaimer } from "../custom-components/Disclaimer";
 import { InlineWidget } from "../components/InlineWidget";
-import { disclaimerSnippet } from "../snippets";
 
 export const content = `
 This disclaimer component can be used as a discrete notice at the beginning of the conversation. For example, they can allow users to opt in for tracking.
 
 ~~~js
-${disclaimerSnippet}
+const Disclaimer = () => {
+  const [status, setStatus] = useState("pending");
+  return html\`
+    <div className="chat-disclaimer">
+      \${status === "pending" &&
+        html\`
+          <p>
+            In order to enhance your experience in this chat, we would like to
+            temporarily store personal data according to our${" "}
+            <a>privacy policy</a>.
+          </p>
+          <button
+            onClick=\${() => {
+              setStatus("accepted");
+            }}
+          >
+            Accept
+          </button>
+          <button
+            onClick=\${() => {
+              setStatus("denied");
+            }}
+          >
+            Deny
+          </button>
+        \`}
+      \${status === "accepted" &&
+        html\`
+          <p>
+            As requested, we might store certain personal or device identifiers
+            to enhance your experience on this chat.
+          </p>
+        \`}
+      \${status === "denied" &&
+        html\`
+          <p>
+            As requested, we will not store personal information in this chat.
+          </p>
+        \`}
+    </div>
+  \`;
+};
 ~~~
 `;
 
-// initial eslint integration
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export const WebWidgetComponentsDisclaimer = () => {
+export const WebWidgetComponentsDisclaimer = (): JSX.Element => {
   return (
     <>
       <PageTitle pretitle="Web widget components" title="Disclaimer" />
