+++ packages/website/src/content/03-06-datepicker.tsx
@@ -3,19 +3,44 @@ import { PageTitle } from "../components/PageTitle";
 import { PageContent } from "../components/PageContent";
 import { DatePicker } from "../custom-components/DatePicker";
 import { InlineWidget, type Item } from "../components/InlineWidget";
-import { datePickerSnippet } from "../snippets";
 
 export const content = `
 This component allows for the user to select a date. For example, the user could select a date of birth, or a date of an appointment. The date picker is a text input that opens a calendar when clicked. The user can then select a date that they are looking for flight tickets.
 
 ~~~js
-${datePickerSnippet}
+export const DatePicker = ({ submitted, onSubmit }) => {
+  const [datepicker, setDatepicker] = useState(null);
+
+  useEffect(() => {
+    const elem = document.getElementById("datepicker");
+    const datepicker = new window.Datepicker(elem, {
+      weekStart: 1, // Monday
+    });
+    datepicker.setDate(new Date());
+    setDatepicker(datepicker);
+  }, []);
+
+  return html\`
+    <form
+      className="chat-datepicker"
+      onSubmit=\${(ev) => {
+        ev.preventDefault();
+
+        const date = datepicker.getDate();
+        onSubmit(date);
+
+        datepicker.setDate({ clear: true });
+      }}
+    >
+      <div id="datepicker"></div>
+      <button type="submit" disabled=\${submitted}>Submit</button>
+    </form>
+  \`;
+};
 ~~~
 `;
 
-// initial eslint integration
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export const WebWidgetComponentsDatePicker = () => {
+export const WebWidgetComponentsDatePicker = (): JSX.Element => {
   const [submitted, setSubmitted] = useState(false);
   const [date, setDate] = useState<Date | null>(null);
