+++ packages/website/src/content/03-08-address-input.tsx
@@ -2,22 +2,108 @@ import React, { useState } from "react";
 import { PageTitle } from "../components/PageTitle";
 import { PageContent } from "../components/PageContent";
 import { InlineWidget, type Item } from "../components/InlineWidget";
-import { addressInputSnippet, mapSnippet } from "../snippets";
 import AddressInput from "../custom-components/Address";
 
 export const content = `
 The address input component is used to collect a user's address. It uses Google Maps to autocomplete the address.
 
 ~~~js
-${mapSnippet}
+const Map = ({ lat, lng, className }) => {
+  const mapRef = useRef(null);
 
-${addressInputSnippet}
+  useEffect(() => {
+    if (mapRef.current) {
+      const map = new google.maps.Map(mapRef.current, {
+        center: { lat, lng },
+        zoom: 15,
+        zoomControl: false,
+        mapTypeControl: false,
+        scaleControl: false,
+        streetViewControl: false,
+        rotateControl: false,
+        fullscreenControl: true,
+      });
+
+      new google.maps.Marker({
+        map: map,
+        position: { lat, lng },
+      });
+    }
+  }, [lat, lng]);
+
+  return html\`<div className=\${className} ref=\${mapRef}></div>\`;
+};
+
+
+const AddressInput = ({ onAddressChange, address, onSubmit, submitted }) => {
+  const [coordinates, setCoordinates] = useState(null);
+  const textareaRef = useRef(null);
+  const [isGoogleMapsLoaded, setIsGoogleMapsLoaded] = useState(false);
+
+  useEffect(() => {
+    const loadGoogleMapsScript = () => {
+      if (window.google) {
+        return;
+      }
+
+      const script = document.createElement("script");
+      script.type = "text/javascript";
+      script.src = 'https://maps.googleapis.com/maps/api/js?key=' + import.meta.env.VITE_GOOGLE_MAPS_API_KEY + '&libraries=places';
+      script.async = true;
+      script.defer = true;
+      script.onload = () => setIsGoogleMapsLoaded(true);
+      document.head.appendChild(script);
+    };
+
+    loadGoogleMapsScript();
+  }, []);
+
+  useEffect(() => {
+    const initializeAutocomplete = () => {
+      if (!window.google || !textareaRef.current) return;
+
+      const autocomplete = new google.maps.places.Autocomplete(
+        textareaRef.current,
+        { types: ['address'] }
+      );
+
+      autocomplete.addListener('place_changed', () => {
+        const place = autocomplete.getPlace();
+        // ... rest of the logic
+      });
+    };
+
+    initializeAutocomplete();
+  }, [isGoogleMapsLoaded]);
+
+  return html\`
+    <form
+      className="address-container"
+      onSubmit=\${(event) => {
+        event.preventDefault();
+        onSubmit();
+      }}
+    >
+      <textarea
+        ref=\${textareaRef}
+        value=\${address}
+        onChange=\${(e) => onAddressChange(e.target.value)}
+        placeholder="Enter address here"
+        rows={5}
+        className="address-textarea"
+      />
+      \${coordinates ?
+        '<Map className="map-container" lat=' + coordinates.lat + ' lng=' + coordinates.lng + ' />' :
+        '<div className="map-placeholder"></div>'
+      }
+      <button disabled=\${!coordinates || submitted} type="submit">Submit</button>
+    </form>
+  \`;
+};
 ~~~
 `;
 
-// initial eslint integration
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export const WebWidgetComponentsAddressInput = () => {
+export const WebWidgetComponentsAddressInput = (): JSX.Element => {
   const [formattedAddress, setFormattedAddress] = useState<string>("");
   const [submitted, setSubmitted] = useState(false);
 
@@ -28,9 +114,13 @@ export const WebWidgetComponentsAddressInput = () => {
         element: (
           <AddressInput
             address={formattedAddress}
-            onAddressChange={(address: string) => { setFormattedAddress(address); }}
+            onAddressChange={(address: string) => {
+              setFormattedAddress(address);
+            }}
             submitted={submitted}
-            onSubmit={() => { setSubmitted(true); }}
+            onSubmit={() => {
+              setSubmitted(true);
+            }}
           />
         ),
       },
