+++ packages/website/src/content/03-10-secure-input.tsx
@@ -13,12 +13,8 @@ ${disclaimerSnippet}
 ~~~
 `;
 
-// initial eslint integration
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export const WebWidgetComponentsSecureInput = () => {
-  // initial eslint integration
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [submitted, setSubmitted] = useState(false);
+export const WebWidgetComponentsSecureInput = (): JSX.Element => {
+  const [submitted] = useState(false);
 
   const items: Item[][] = [
     [
